// application/dto/main/types.gen.ts
export type Event = {
    /**
     * Event Type Id
     */
    event_type_id?: number | null;
    /**
     * Title
     */
    title?: string | null;
    /**
     * Description
     */
    description?: string | null;
    /**
     * Date From
     */
    date_from?: string | null;
    /**
     * Date To
     */
    date_to?: string | null;
    /**
     * Is Active
     */
    is_active?: boolean;
    /**
     * Id
     */
    id: number;
    event_type?: EventType | null;
    /**
     * User Id
     */
    user_id?: string | null;
    /**
     * Duration
     */
    duration?: number | null;
    /**
     * Updated On
     */
    updated_on?: string;
    /**
     * Created On
     */
    created_on?: string;
};

/**
 * EventCreate
 */
export type EventCreate = {
    /**
     * Event Type Id
     */
    event_type_id?: number | null;
    /**
     * Title
     */
    title?: string | null;
    /**
     * Description
     */
    description?: string | null;
    /**
     * Date From
     */
    date_from?: string | null;
    /**
     * Date To
     */
    date_to?: string | null;
};
// ...
export type EventUpdate = {
    /**
     * Event Type Id
     */
    event_type_id?: number | null;
    /**
     * Title
     */
    title?: string | null;
    /**
     * Description
     */
    description?: string | null;
    /**
     * Date From
     */
    date_from?: string | null;
    /**
     * Date To
     */
    date_to?: string | null;
    /**
     * Is Active
     */
    is_active?: boolean | null;
};
// ...
export type PaginationSchemaEvent = {
    /**
     * Total
     */
    total: number;
    /**
     * Page
     */
    page: number;
    /**
     * Total Pages
     */
    total_pages: number;
    /**
     * Items
     */
    items: Array<Event>;
};
// ...
export type ReadEventsApiV1EventsGetResponses = {
    /**
     * Successful Response
     */
    200: PaginationSchemaEvent;
};

export type ReadEventsApiV1EventsGetResponse = ReadEventsApiV1EventsGetResponses[keyof ReadEventsApiV1EventsGetResponses];

export type CreateNewEventApiV1EventsPostData = {
    body: EventCreate;
    path?: never;
    query?: never;
    url: '/api/v1/events/';
};

export type CreateNewEventApiV1EventsPostErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type CreateNewEventApiV1EventsPostError = CreateNewEventApiV1EventsPostErrors[keyof CreateNewEventApiV1EventsPostErrors];

export type CreateNewEventApiV1EventsPostResponses = {
    /**
     * Successful Response
     */
    200: Event;
};

export type CreateNewEventApiV1EventsPostResponse = CreateNewEventApiV1EventsPostResponses[keyof CreateNewEventApiV1EventsPostResponses];
// ...
export type DeleteExistingEventApiV1EventsEventIdDeleteResponses = {
    /**
     * Successful Response
     */
    200: Event;
};

export type DeleteExistingEventApiV1EventsEventIdDeleteResponse = DeleteExistingEventApiV1EventsEventIdDeleteResponses[keyof DeleteExistingEventApiV1EventsEventIdDeleteResponses];
// ...
export type ReadEventApiV1EventsEventIdGetResponses = {
    /**
     * Successful Response
     */
    200: Event;
};

export type ReadEventApiV1EventsEventIdGetResponse = ReadEventApiV1EventsEventIdGetResponses[keyof ReadEventApiV1EventsEventIdGetResponses];
// ...
export type UpdateExistingEventApiV1EventsEventIdPutResponses = {
    /**
     * Successful Response
     */
    200: Event;
};

export type UpdateExistingEventApiV1EventsEventIdPutResponse = UpdateExistingEventApiV1EventsEventIdPutResponses[keyof UpdateExistingEventApiV1EventsEventIdPutResponses];